Add script to determine if the kernel has the needed options.
Exits with a non-zero exit code if the system's kernel doesn't support
eBPF collector requirements.

This only issues warnings if the kernel version is too old, because
RHEL/CentOS are questionable and use ancient kernel version numbers even
though they have very new features.

THis first checks if we're running on Linux, and bails early if we
aren't.

Once checking the version and possibly issuing a warning, it then looks
for the configuration in the following places, using the first it finds:

* `/proc/config.gz` (we also try to load the module that provides this).
* `/lib/modules/${KERNEL_VERSION}/sources/.config`
* `/boot/config-${KERNEL_VERSION}`

If it can find the config, it uses grep to check for all the required
options being enabled.

if it can't find the config, it assumes things will work, and exits with
0.