oomkill: kprobe to a sig that makes sense for the kernel vsn.
on newer versions, it makes sense to use out_of_memory, because
it seems to be quite stable and also the entrypoint for OOM
killing.

on older versions, that function doesnt give us what we want. so
we use oom_kill_process, which doesnt change for the kernel
versions we support in that older range.