Switch from xz to gz for temporary artifacts.
The artifacts produced by the Makefile are only used for development and
for internal transfer of the build artifacts within the CI jobs, so the
choice of compression doesn’t really matter much here.

XZ seems to be having issues inside the GHA runners for CI, so we need
to use something else. Gzip is the most logical choice here since it
should be everywhere.